refactor(frontend): extract shared Case-Incident custom field components
Move the customFieldDefinitionsForEntityType query, the CustomFieldDef
type, getCustomFieldSetting helper and the CaseIncidentCustomFieldInput
renderer out of CaseIncidentCreation.tsx into a new shared module
(CaseIncidentCustomFields.tsx), so the edition overview can reuse them
instead of duplicating ~150 lines.

The markdown field type now renders via the fully-controlled
MarkdownFieldBase (value/onValueChange) instead of a Formik <Field>,
making CaseIncidentCustomFieldInput usable outside of a Formik context
(needed for the per-field blur-submit edition pattern).
