+++ opencti-platform/opencti-front/src/private/components/cases/case_incidents/CaseIncidentCreation.tsx
@@ -5,12 +5,6 @@ import AuthorizedMembersField from '@components/common/form/AuthorizedMembersFie
 import AccordionDetails from '@mui/material/AccordionDetails';
 import Typography from '@mui/material/Typography';
 import Divider from '@mui/material/Divider';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import MenuItem from '@mui/material/MenuItem';
-import MuiAutocomplete from '@mui/material/Autocomplete';
-import Chip from '@mui/material/Chip';
-import MuiTextField from '@mui/material/TextField';
-import Switch from '@mui/material/Switch';
 import { Field, Form, Formik } from 'formik';
 import { FormikConfig } from 'formik/dist/types';
 import { FunctionComponent, useState } from 'react';
@@ -23,7 +17,6 @@ import { Accordion, AccordionSummary } from '../../../../components/Accordion';
 import FormButtonContainer from '../../../../components/common/form/FormButtonContainer';
 import CreateEntityControlledDial from '../../../../components/CreateEntityControlledDial';
 import DateTimePickerField from '../../../../components/DateTimePickerField';
-import DatePicker from '../../../../components/common/input/DatePicker';
 import MarkdownField from '../../../../components/fields/markdownField/MarkdownField';
 import RichTextField from '../../../../components/fields/RichTextField';
 import { useFormatter } from '../../../../components/i18n';
@@ -48,143 +41,8 @@ import ObjectMarkingField from '../../common/form/ObjectMarkingField';
 import ObjectParticipantField from '../../common/form/ObjectParticipantField';
 import OpenVocabField from '../../common/form/OpenVocabField';
 import { CaseIncidentAddInput, CaseIncidentCreationCaseMutation } from './__generated__/CaseIncidentCreationCaseMutation.graphql';
-import {
-  CaseIncidentCreationCustomFieldDefinitionsQuery,
-  CaseIncidentCreationCustomFieldDefinitionsQuery$data,
-} from './__generated__/CaseIncidentCreationCustomFieldDefinitionsQuery.graphql';
-
-const customFieldDefinitionsForEntityTypeQuery = graphql`
-  query CaseIncidentCreationCustomFieldDefinitionsQuery($entityType: String!) {
-    customFieldDefinitionsForEntityType(entityType: $entityType) {
-      edges {
-        node {
-          id
-          name
-          label
-          field_type
-          min_value
-          max_value
-          select_options
-          entity_type_settings {
-            entity_type
-            mandatory
-            default_value
-          }
-        }
-      }
-    }
-  }
-`;
-
-type CustomFieldDef = NonNullable<NonNullable<CaseIncidentCreationCustomFieldDefinitionsQuery$data['customFieldDefinitionsForEntityType']>['edges']>[number]['node'];
-
-const getCustomFieldSetting = (definition: CustomFieldDef, entityType: string) => (definition.entity_type_settings ?? []).find((setting) => setting.entity_type === entityType);
-
-// Renders the appropriate Formik-less input for a custom field definition; value/onChange are wired to Formik state by the caller.
-const CaseIncidentCustomFieldInput: FunctionComponent<{
-  definition: CustomFieldDef;
-  mandatory: boolean;
-  value: string | boolean | string[];
-  onChange: (val: string | boolean | string[]) => void;
-}> = ({ definition, mandatory, value, onChange }) => {
-  const { t_i18n } = useFormatter();
-  const label = `${definition.label}${mandatory ? ' *' : ''}`;
-
-  if (definition.field_type === 'boolean') {
-    return (
-      <FormControlLabel
-        style={fieldSpacingContainerStyle}
-        control={(
-          <Switch
-            checked={value === true}
-            onChange={(_, checked) => onChange(checked)}
-          />
-        )}
-        label={label}
-      />
-    );
-  }
-  if (definition.field_type === 'date') {
-    return (
-      <DatePicker
-        value={value ? new Date(String(value)) : null}
-        onChange={(date) => onChange(date ? date.toISOString() : '')}
-        label={label}
-        slotProps={{ textField: { variant: 'standard', fullWidth: true, style: fieldSpacingContainerStyle } }}
-      />
-    );
-  }
-  if (definition.field_type === 'select' && definition.select_options) {
-    return (
-      <MuiTextField
-        select
-        fullWidth
-        variant="standard"
-        label={label}
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-        style={fieldSpacingContainerStyle}
-      >
-        <MenuItem value=""><em>{t_i18n('None')}</em></MenuItem>
-        {definition.select_options.map((opt) => (
-          <MenuItem key={opt} value={opt}>{opt}</MenuItem>
-        ))}
-      </MuiTextField>
-    );
-  }
-  if (definition.field_type === 'multi_select' && definition.select_options) {
-    const selected = Array.isArray(value) ? value : [];
-    return (
-      <MuiAutocomplete
-        multiple
-        options={definition.select_options}
-        value={selected}
-        onChange={(_, newValue) => onChange(newValue)}
-        renderTags={(tagValue, getTagProps) => tagValue.map((option: string, index: number) => (
-          <Chip label={option} {...getTagProps({ index })} key={option} />
-        ))}
-        renderInput={(params) => (
-          <MuiTextField
-            {...params}
-            variant="standard"
-            label={label}
-            style={fieldSpacingContainerStyle}
-          />
-        )}
-      />
-    );
-  }
-  if (definition.field_type === 'markdown') {
-    return (
-      <Field
-        component={MarkdownField}
-        name={`customFields.${definition.id}`}
-        label={label}
-        required={mandatory}
-        fullWidth
-        multiline
-        rows="4"
-        style={fieldSpacingContainerStyle}
-      />
-    );
-  }
-  return (
-    <MuiTextField
-      fullWidth
-      variant="standard"
-      label={label}
-      value={value}
-      type={definition.field_type === 'integer' ? 'number' : 'text'}
-      inputProps={
-        definition.field_type === 'integer'
-          ? { min: definition.min_value ?? undefined, max: definition.max_value ?? undefined }
-          : undefined
-      }
-      onChange={(e) => onChange(e.target.value)}
-      style={fieldSpacingContainerStyle}
-    />
-  );
-};
+import { CaseIncidentCustomFieldsQuery } from './__generated__/CaseIncidentCustomFieldsQuery.graphql';
+import { CaseIncidentCustomFieldInput, customFieldDefinitionsForEntityTypeQuery, CustomFieldDef, getCustomFieldSetting } from './CaseIncidentCustomFields';
 
 const caseIncidentMutation = graphql`
   mutation CaseIncidentCreationCaseMutation($input: CaseIncidentAddInput!) {
@@ -264,7 +122,7 @@ export const CaseIncidentCreationForm: FunctionComponent<IncidentFormProps> = ({
   const { mandatoryAttributes } = useIsMandatoryAttribute(
     CASE_INCIDENT_TYPE,
   );
-  const customFieldData = useLazyLoadQuery<CaseIncidentCreationCustomFieldDefinitionsQuery>(
+  const customFieldData = useLazyLoadQuery<CaseIncidentCustomFieldsQuery>(
     customFieldDefinitionsForEntityTypeQuery,
     { entityType: CASE_INCIDENT_TYPE },
   );
